+++ src/test/java/org/folio/rest/impl/LocationUnitInstitutionIT.java
@@ -5,6 +5,7 @@ import static org.folio.HttpStatus.HTTP_BAD_REQUEST;
 import static org.folio.HttpStatus.HTTP_INTERNAL_SERVER_ERROR;
 import static org.folio.HttpStatus.HTTP_NOT_FOUND;
 import static org.folio.HttpStatus.HTTP_NO_CONTENT;
+import static org.folio.HttpStatus.HTTP_OK;
 import static org.folio.HttpStatus.HTTP_UNPROCESSABLE_ENTITY;
 import static org.folio.rest.impl.LocationUnitApi.CAMPUS_TABLE;
 import static org.folio.services.locationunit.InstitutionService.INSTITUTION_TABLE;
@@ -16,11 +17,14 @@ import io.vertx.core.Future;
 import io.vertx.core.Vertx;
 import io.vertx.junit5.VertxExtension;
 import io.vertx.junit5.VertxTestContext;
+import java.util.ArrayList;
 import java.util.List;
+import java.util.Map;
 import java.util.Objects;
 import java.util.UUID;
 import java.util.function.Function;
 import java.util.function.UnaryOperator;
+import org.assertj.core.api.InstanceOfAssertFactories;
 import org.folio.rest.jaxrs.model.Error;
 import org.folio.rest.jaxrs.model.Errors;
 import org.folio.rest.jaxrs.model.Loccamp;
@@ -112,6 +116,49 @@ class LocationUnitInstitutionIT
       .compose(rows -> postgresClient.delete(INSTITUTION_TABLE, (CQLWrapper) null))
       .onFailure(ctx::failNow)
       .onComplete(event -> ctx.completeNow());
+  }
+
+  @Test
+  void getCollection_shouldReturnRecordCollectionBasedOnIncludeShadowQueryParam(Vertx vertx, VertxTestContext ctx) {
+    var client = vertx.createHttpClient();
+    var postgresClient = PostgresClient.getInstance(vertx, TENANT_ID);
+
+    var nonShadowInstitution = sampleRecord();
+    var shadowInstitution = sampleRecord()
+      .withIsShadow(true)
+      .withName("test-shadow-institution")
+      .withCode("shadow");
+    var institutionByQueryParam = Map.of(false, nonShadowInstitution, true, shadowInstitution);
+
+    Future.all(
+        postgresClient.save(referenceTable(), nonShadowInstitution),
+        postgresClient.save(referenceTable(), shadowInstitution)
+      )
+      .compose(s -> {
+        List<Future<TestResponse>> futures = new ArrayList<>();
+        for (boolean param : institutionByQueryParam.keySet()) {
+          var responseFuture = doGet(client, resourceUrl() + "?includeShadow=" + param)
+            .onComplete(verifyStatus(ctx, HTTP_OK))
+            .andThen(ctx.succeeding(response -> ctx.verify(() -> {
+              var collectionUnits = response.bodyAsClass(Locinsts.class);
+              assertThat(collectionUnits)
+                .as("verify collection for query param and value: includeShadow=" + param)
+                .isNotNull()
+                .hasFieldOrPropertyWithValue("totalRecords", 1)
+                .extracting(Locinsts::getLocinsts).asInstanceOf(InstanceOfAssertFactories.COLLECTION)
+                .hasSize(1);
+
+              var collectionRecord = collectionUnits.getLocinsts().getFirst();
+              verifyRecordFields(collectionRecord, institutionByQueryParam.get(param),
+                List.of(Locinst::getName, Locinst::getIsShadow),
+                "verify collection's record for query param: ?includeShadow=" + param);
+            })));
+          futures.add(responseFuture);
+        }
+        return Future.all(futures);
+      })
+      .onFailure(ctx::failNow)
+      .onSuccess(event -> ctx.completeNow());
   }
 
   @Test